pyproject.toml
[tool.poetry]
name = "seedx_support_backend"
version = "0.1.0"
description = "seedx support backend"
authors = ["Olyad Temesgen <olexyit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.1,<0.35.0"
sqlalchemy = ">=2.0.40,<3.0.0"
pydantic = ">=2.11.3,<3.0.0"
python-jose = ">=3.4.0,<4.0.0"
passlib = ">=1.7.4,<2.0.0"
python-multipart = ">=0.0.20,<0.0.21"
aiohttp = ">=3.11.16,<4.0.0"
injector = ">=0.22.0,<0.23.0"
pymongo = ">=4.12.0,<5.0.0"
pyjwt = ">=2.10.1,<3.0.0"
ulid-py = ">=1.1.0,<2.0.0"
cattrs = ">=24.1.3,<25.0.0"
python-dateutil = ">=2.9.0.post0,<3.0.0"
httpx = ">=0.28.1,<0.29.0"
alembic = ">=1.15.2,<2.0.0"
pydantic-settings = ">=2.9.1,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
email-validator = ">=2.2.0,<3.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
